+++ server/tests/test_google_drive_event_normalizer.py
@@ -65,3 +65,15 @@ def test_normalize_drive_activity_basic():
 
 
 
+
+
+
+
+
+
+
+
+
+
+
+